golf-coach-chatbot/backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
firebase-admin==6.4.0
google-cloud-firestore==2.14.0
openai>=1.40.0
python-dotenv==1.0.0
werkzeug==3.0.1
httpx>=0.27.0
opencv-python>=4.8.0
numpy>=1.24.0

